Refactoring! Doesn't everybody love that word?
In this release:

Version number bumped to 0.9.6.0
The project now targets the full .NET 4 framework, not just the client profile.

DuneCommands/*Command.vb:
1) Commands now make extensive use of constants in Constants.vb
2) GetQueryString() now builds on a more generic GetQuery() method, taking advantage of HttpUtility methods.

FirmwareInfo.vb: added a firmware checker

IRemoteControl.vb: removed the need for a custom code attribute, codes are now stored as hexadecimal enum values.

RemoteControlCommand.vb: button codes are now retrieved using a bitconverter instead of reflection.

UrlConverter.vb:
1) Option Compare is now Text.
2) Does no longer encode the URLs (instead this is done in the GetQuery() call on StartPlaybackCommand).

Constants.vb: More constants! Yess! (Oh and also in-line documentation)

CommandResult.vb: The constructor is now for internal use only. To get a command result, use the FromCommand() or FromCommandAsync() calls. That's right, we're going on the multithreaded highway! Yes!

Dune.vb:
1) Added information about the device's network interface card.
2) Added async commands.
3) Added blackscreen and hide OSD (write-only) properties.
4) A bunch of properties have been upgrades from read-only to get or set.

Every change not listed here is either not very important or concerns a bugfix. Unless it's not a bugfix, which means it was important... shut up.